Complete Pie Changes *Bugs *Fingerprint,VoLTE(WIP),Camera

+++ camera/QCamera2/HAL/QCameraPostProc.cpp
@@ -2214,7 +2214,7 @@ void *QCameraPostProcessor::dataSaveRoutine(void *data)
                         if ((ssize_t)job_data->out_data.buf_filled_len != written_len) {
                             ALOGE("%s: Failed save complete data %zd bytes "
                                   "written instead of %d bytes!",
-                                  __func__, (long int)written_len,
+                                  __func__, written_len,
                                   job_data->out_data.buf_filled_len);
                         } else {
                             CDBG_HIGH("%s: written number of bytes %zd\n",
